separate private and public handlers for topic in different files
