+++ api/api.go
@@ -66,7 +66,7 @@ func Setup(ctx context.Context, cfg *config.Config, router *mux.Router, dataStor
 
 // enablePublicEndpoints register only the public GET endpoints.
 func (api *API) enablePublicEndpoints(ctx context.Context) {
-	api.get("/topics", api.getTopicsListPublicHandler)
+	api.get("/topics", api.getRootTopicsPublicHandler)
 	api.get("/topics/{id}", api.getTopicPublicHandler)
 	api.get("/topics/{id}/subtopics", api.getSubtopicsPublicHandler)
 	api.get("/topics/{id}/content", api.getContentPublicHandler)
@@ -97,7 +97,7 @@ func (api *API) enablePrivateTopicEndpoints(ctx context.Context) {
 	api.get(
 		"/topics",
 		api.isAuthenticated(
-			api.isAuthorised(readPermission, api.getTopicsListPrivateHandler)),
+			api.isAuthorised(readPermission, api.getRootTopicsPrivateHandler)),
 	)
 
 	api.get(
